Login feature Added
I added a login feature to Rate, Cancel, and Change Windows. It works to an extent,  just sometimes doesn't function when you login. We should probably fix that and just edit the code for that. Some things that might have to get done is making the GUI components of each window and just fixing overall bugs. Feel free to edit/ improve or organize any of the functions or code segments. Sorry if the code is messy btw, for now I was just workin on getting things to work but in the future, we'll improve the code. Maybe