Update Main.java
I added a table that imports all the data in hotel room in a method called table in the SearchWindow.class. I also added a function called ReturnCollection in the HotelRoom.class which is used to make other classes get the data from HotelRoom. I removed Hotel.LoadData method (as well as Customer.LoadData) and found it more ideal to move it to the Main class, so it processes the info before .